+++ parser/parser.go
@@ -3,7 +3,6 @@ package parser
 import (
 	"fmt"
 	"reflect"
-	"strings"
 )
 
 func Plus(s Stream, p Parser) (Stream, []Value) {
@@ -32,53 +31,6 @@ func Repeat(s Stream, min, max int, p Parser) (Stream, []Value) {
 		vs = append(vs, p)
 	}
 	return s, vs
-}
-
-type SyntaxError interface {
-	error
-	AddContext(SyntaxErrorContext)
-}
-
-type syntaxError struct {
-	Err   error
-	Stack []SyntaxErrorContext
-}
-
-func (se *syntaxError) AddContext(c SyntaxErrorContext) {
-	se.Stack = append(se.Stack, c)
-}
-
-func NewSyntaxError(c SyntaxErrorContext) SyntaxError {
-	se := new(syntaxError)
-	se.Stack = append(se.Stack, c)
-	return se
-}
-
-type SyntaxErrorContext struct {
-	Stream Stream
-	Parser Parser
-	Err    error
-}
-
-func (se *syntaxError) Error() string {
-	var ss []string
-	for _, c := range se.Stack {
-		if c.Err != nil {
-			ss = append(ss, c.Err.Error())
-		}
-		if c.Parser != nil {
-			ss = append(ss, fmt.Sprintf("while parsing %s", ParserName(c.Parser)))
-		}
-		if c.Stream != nil {
-			ss = append(ss, fmt.Sprintf("starting with %q at %s", c.Stream.Token(), c.Stream.Position()))
-		}
-	}
-	s := strings.Join(ss, " ")
-	ret := "syntax error"
-	if s != "" {
-		ret += ": " + s
-	}
-	return ret
 }
 
 func ParseErr(s Stream, p Parser) (_s Stream, err SyntaxError) {
